+++ CHANGELOG.md
@@ -17,6 +17,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.6.4] - 2025-01-13
+
+### Fixed
+
+- Chat Input regressions
+- Navigation in Project Manager
+- Propagating errors from LLM
+
+
 ## [0.6.3] - 2025-01-12
 
 ### Fixed